Time: 0 ms (100.00%), Space: 6 MB (56.03%) - LeetHub

+++ 0556-next-greater-element-iii/0556-next-greater-element-iii.cpp
@@ -35,6 +35,6 @@ public:
         if (res > INT_MAX) {
             return -1;
         }
-        return res;
+        return static_cast<int>(res);
     }
 };
